Python 2.x -> 3.x
* Caught numeric issue

+++ code/crawler.py
@@ -122,10 +122,10 @@ class CrawlingRobotEnvironment(environment.Environment):
         ## Also call self.crawlingRobot.setAngles()
         ## to the initial arm and hand angle
 
-        armState = self.nArmStates/2
-        handState = self.nHandStates/2
+        armState = int(self.nArmStates/2)
+        handState = int(self.nHandStates/2)
         self.state = armState,handState
-        self.crawlingRobot.setAngles(self.armBuckets[int(armState)],self.handBuckets[int(handState)])
+        self.crawlingRobot.setAngles(self.armBuckets[armState],self.handBuckets[handState])
         self.crawlingRobot.positions = [20,self.crawlingRobot.getRobotPosition()[0]]
